refactor: move drop_down_date_dialog to widget folder
